move event listen from page status alert to page
